mcp_server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "MCP server for semantic layer SQL metadata retrieval."
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.14.4",
    "mcp>=1.26.0",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "neo4j>=6.1.0",
    "pandas>=3.0.0",
]

[project.scripts]
mcp-server = "mcp_server.server:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server"]
